Extract `connections` module
